Implemented a massive upgrade:
Implemented a custom hook called useAxios to intercept request and response objects. If the access token expires, the refresh token will be sent to the backend programmatically to renew the access token via this hook.

Implemented a user profile component for testing purposes. I will work on improving it further.